Additional initializers; Updated static var references; Moved public extensions to "+Computed" source files;
