+++ Sources/Swift2D/Point+Computed.swift
@@ -1,0 +1,6 @@
+// MARK: - Static References
+public extension Point {
+    static let zero: Point = Point(x: 0, y: 0)
+    static let nan: Point = Point(x: Float.nan, y: Float.nan)
+    static let infinite: Point = Point(x: Float.infinity, y: Float.infinity)
+}
