+++ Sources/Swift2D/Point.swift
@@ -1,5 +1,3 @@
-import Foundation
-
 /// The representation of a single point in a two-dimensional plane.
 public struct Point {
     public var x: Float
@@ -19,6 +17,11 @@ public struct Point {
         self.x = Float(x)
         self.y = Float(y)
     }
+    
+    public init(x: Double, y: Double) {
+        self.x = Float(x)
+        self.y = Float(y)
+    }
 }
 
 // MARK: - CustomStringConvertible
@@ -30,10 +33,19 @@ extension Point: CustomStringConvertible {
 
 // MARK: - Equatable
 extension Point: Equatable {
+    public static func ==(lhs: Point, rhs: Point) -> Bool {
+        if (lhs.x.isNaN && rhs.x.isNaN) && (lhs.y.isNaN && rhs.y.isNaN) {
+            return true
+        }
+        
+        return lhs.x == rhs.x && lhs.y == rhs.y
+    }
 }
 
-// MARK: - Static References
-public extension Point {
-    static let nan: Point = Point(x: .nan, y: .nan)
-    static let zero: Point = Point(x: 0.0, y: 0.0)
+// MARK: - Hashable
+extension Point: Hashable {
+}
+
+// MARK: - Codable
+extension Point: Codable {
 }
