+++ Sources/Swift2D/Size+Computed.swift
@@ -1,0 +1,34 @@
+// MARK: - Static References
+public extension Size {
+    static let zero: Size = Size(width: 0.0, height: 0.0)
+    static let nan: Size = Size(width: Float.nan, height: Float.nan)
+    static let infinite: Size = Size(width: Float.infinity, height: Float.infinity)
+}
+
+// MARK: - Instance Functionality
+public extension Size {
+    /// The horizontal radius (½ of `width`)
+    var xRadius: Float {
+        return abs(width) / 2.0
+    }
+    
+    /// The vertical radius (½ of `height`)
+    var yRadius: Float {
+        return abs(height) / 2.0
+    }
+    
+    /// The largest radius, out of `xRadius` & `yRadius`.
+    var maxRadius: Float {
+        return max(xRadius, yRadius)
+    }
+    
+    /// The smallest radius, out of `xRadius` & `yRadius`.
+    var minRadius: Float {
+        return min(xRadius, yRadius)
+    }
+    
+    /// The `Point` at which the `xRadius` & `yRadius` intersect.
+    var center: Point {
+        return Point(x: xRadius, y: yRadius)
+    }
+}
